GCP/sample/requirements.txt
protobuf==4.21.12
mysql-connector-python==8.2.0